refactor: move kit.Ptr to internal
