+++ proto/value_marshal_test.go
@@ -14,7 +14,7 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	"github.com/muktihari/fit/kit"
+	"github.com/muktihari/fit/internal/kit"
 )
 
 func TestValueMarshalAppend(t *testing.T) {
